// src/lib/aisc/types.ts
export type AiscShapeType = string;

export type AiscShape = {
  shape: string;
  type: AiscShapeType;
  W: number;
  A: number;
  d: number;
  h: number;
  bf: number;
  tf: number;
  tw: number;
  Ix: number;
  Iy: number;
  Zx: number;
  Sx: number;
  Zy: number;
  rx: number;
  ry: number;
  bf_2tf: number;
  h_tw: number;
};